Move code not related to actual cases to a helper #42
so that we can more easily scale and keep the test cases readable
