+++ tests/dataProvider.test.ts
@@ -1,15 +1,4 @@
-import { fetchUtils } from 'ra-core';
-import raPostgrestProvider from '../src/index';
-
-type Case = {
-    test: string;
-    method: string;
-    resource: string;
-    params: Record<string, any>;
-    responseHeaders?: Record<string, string>;
-    expectedUrl: string;
-    expectedHeaders: Record<string, string>;
-};
+import { makeTestFromCase, Case } from './helper';
 
 const cases: Case[] = [
     {
@@ -51,59 +40,5 @@ const cases: Case[] = [
 ];
 
 describe('urlBuilder', () => {
-    cases.forEach(
-        ({
-            test,
-            method,
-            resource,
-            params,
-            responseHeaders,
-            expectedUrl,
-            expectedHeaders,
-        }) => {
-            it(`${method} should build correct url for a ${test}`, async () => {
-                const { httpClient, dataPovider } = createDataProviderMock(
-                    200,
-                    '',
-                    [],
-                    responseHeaders
-                );
-
-                await dataPovider[method](resource, params);
-
-                expect(httpClient).toHaveBeenCalledWith(
-                    `${BASE_URL}${expectedUrl}`,
-                    makeHeaderExpectation(expectedHeaders)
-                );
-            });
-        }
-    );
+    cases.forEach(makeTestFromCase);
 });
-
-type HTTPClientMock = typeof fetchUtils.fetchJson;
-const BASE_URL = 'http://localhost:3000';
-
-function createDataProviderMock(
-    expectedStatus: number,
-    expectedBody: string,
-    expectedJSON?: any,
-    expectedHeaders?: Record<string, string>
-) {
-    const httpClient = jest.fn(url =>
-        Promise.resolve({
-            status: expectedStatus,
-            headers: new Headers(expectedHeaders),
-            body: expectedBody,
-            json: expectedJSON,
-        })
-    );
-    const dataPovider = raPostgrestProvider(BASE_URL, httpClient);
-
-    return { httpClient, dataPovider };
-}
-
-function makeHeaderExpectation(headers) {
-    return {
-        headers: new Headers(headers),
-    };
-}
